+++ server-data/resources/[bpt_addons]/bpt_importjob/locales/it.lua
@@ -4,6 +4,11 @@ Locales['it'] = {
   ['cloakroom_prompt'] = 'premi [E] per accedere al guardaroba.',
   ['wear_citizen'] = 'abiti civili',
   ['wear_work'] = 'abiti da lavoro',
+  -- Inventory
+  ['deposit_stock'] = 'Depositare',
+  ['take_stock'] = 'Prendi',
+  ['have_deposited'] = 'Hai depositato',
+  ['quantity_invalid'] = 'Quantità non valida',
   -- garage
   ['spawner_prompt'] = 'premi [E] per accedere al Garage.',
   ["vehicle_spawned"] = "Veicolo ritirato con successo!",
@@ -21,5 +26,8 @@ Locales['it'] = {
   ['no_players_near'] = 'nessun giocatore nelle vicinanze',
   ['boss_actions'] = 'Azioni del boss',
   ['blip_import'] = 'Import/Export',
-  ['import'] = 'import'
+  ['import'] = 'import',
+  -- billing 
+  ['bill_amount'] = "importo della fattura",
+  ['confirm'] = "Conferma"
 }