+++ server-data/resources/[bpt_addons]/bpt_importjob/server/main.lua
@@ -2,6 +2,10 @@ TriggerEvent('esx_society:registerSociety', 'import', 'Import', 'society_import'
     type = 'public'
 })
 
+if Config.MaxInService ~= -1 then
+    TriggerEvent('esx_service:activateService', 'import', Config.MaxInService)
+end
+
 ESX.RegisterServerCallback("bpt_importjob:SpawnVehicle", function(source, cb, model , props)
     local xPlayer = ESX.GetPlayerFromId(source)
 
@@ -18,4 +22,59 @@ ESX.RegisterServerCallback("bpt_importjob:SpawnVehicle", function(source, cb, mo
         TaskWarpPedIntoVehicle(GetPlayerPed(source), vehicle, -1)
     end)
     cb()
+end)
+
+RegisterNetEvent('bpt_importjob:getStockItem')
+AddEventHandler('bpt_importjob:getStockItem', function(itemName, count)
+    local xPlayer = ESX.GetPlayerFromId(source)
+
+    if xPlayer.job.name == 'import' then
+        TriggerEvent('esx_addoninventory:getSharedInventory', 'society_import', function(inventory)
+            local item = inventory.getItem(itemName)
+
+            -- is there enough in the society?
+            if count > 0 and item.count >= count then
+                -- can the player carry the said amount of x item?
+                if xPlayer.canCarryItem(itemName, count) then
+                    inventory.removeItem(itemName, count)
+                    xPlayer.addInventoryItem(itemName, count)
+                    xPlayer.showNotification(_U('have_withdrawn', count, item.label))
+                else
+                    xPlayer.showNotification(_U('player_cannot_hold'))
+                end
+            else
+                xPlayer.showNotification(_U('quantity_invalid'))
+            end
+        end)
+    else
+        print(('[^3WARNING^7] Player ^5%s^7 attempted ^5bpt_importjob:getStockItem^7 (cheating)'):format(source))
+    end
+end)
+
+ESX.RegisterServerCallback('bpt_importjob:getStockItems', function(_, cb)
+    TriggerEvent('esx_addoninventory:getSharedInventory', 'society_import', function(inventory)
+        cb(inventory.items)
+    end)
+end)
+
+RegisterNetEvent('bpt_importjob:putStockItems')
+AddEventHandler('bpt_importjob:putStockItems', function(itemName, count)
+    local xPlayer = ESX.GetPlayerFromId(source)
+	local sourceItem = xPlayer.getInventoryItem(itemName)
+
+    if xPlayer.job.name == 'import' then
+        TriggerEvent('esx_addoninventory:getSharedInventory', 'society_import', function(inventory)
+            local item = inventory.getItem(itemName)
+
+            if sourceItem.count >= count and count > 0 then
+                xPlayer.removeInventoryItem(itemName, count)
+                inventory.addItem(itemName, count)
+                xPlayer.showNotification(_U('have_deposited', count, item.label))
+            else
+                xPlayer.showNotification(_U('quantity_invalid'))
+            end
+        end)
+    else
+        print(('[^3WARNING^7] Player ^5%s^7 attempted ^5bpt_importjob:putStockItems^7 (cheating)'):format(source))
+    end
 end)